requirements.txt
requests~=2.31.0
pandas~=1.5.1
pytz~=2023.2
numpy~=1.24.2
xmltodict~=0.13.0
python-dateutil~=2.8.2
aiohttp~=3.8.4

